gst/msdk: move AVC encoder test classes to lib
